requirements.txt
Flask==2.1.2
Werkzeug==2.1.2
wakeonlan==3.1.0
python-dotenv==0.20.0
